s5e8825: init: Tune Mali GPU interactive gov boosting
We have to be more aggressive because boosting in AOSP doesn't work
like in One UI, and the GPU's performance for rendering the UI is really
poor in the lower frequencies (302 - 507 MHz).

New config:
- hispeed load: 80%
- hispeed freq: 702 MHz

With this it will more often boost up to 702 MHz during light usage
and get rid of those stutters.

Change-Id: I53cf66a243cd2b79173ba0858bdd31e18750a84d
Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>
Signed-off-by: Majaahh <maja.blaszczak@proton.me>